Extract components from PaymentDetails
